task3/requirements.txt
xmltodict
beautifulsoup4
pandas
openpyxl
PyGithub
python-dotenv
sphinx
sphinx_rtd_theme
ghp_import
pytest
sphinx-needs
sphinxcontrib-plantuml